Update style guide XML to new format
Folks should be on Android Studio 1.2 or later, which I think
changed the format of this file.

Note: I didn't really notice this for a long time, because I think when
it upgraded past 1.2, it copied my settings into `~/Library/Preferences/AndroidStudio1.2/`
but instead of keeping a symlink, it copied the XML file into a real file.
I recently went in there and changed it back to a symlink and it now keeps
doing this.