Revamp the style guide formatting, rewrite it, and add a few rules.
Summary:
Maybe I should have done this as separate commits, but the file is
small enough I think it's reasonable to deal with them all in one go.

Major formatting changes:

* We use h3 and h4 rather than h2 and h3, which makes more of the
  style guide fit on a screen at a time, and in general improves
  readability (I think).
* The style guide is now divided into logical sections ("formatting,"
  "libraries," etc.), each with a small number of items in them.  I
  think this improves readability as well.
* I have standardized on 'yes'/'no' text for examples, rather than
  sometimes having 'good'/'bad'.  I like 'yes' or 'no': style guides
  should be opinionated.

Likewise, I reworded a bunch of stuff to say "do this, do that" rather
than just "this is good, that is good."  Other rewording was mostly to
tighten the text.  I also got rid of some obsolete text (such as a
mention of the new package system, which never materialized).  I also
tried to rewrite the 'standard components' section so it doesn't get
out of date quite so quickly.

I added two new rules:

* Only export a single react class.  This came out of the best
  practices doc at
  https://docs.google.com/document/d/1ChtFUao18IyNhaXZ5sE2W-CFuFcYnqlFTyi5gfe6XV0/edit#heading=h.10yidnt3j9ml
* Some discussion about "dumb" and "smart" components.

I also tightened the rule about using propTypes.

Test Plan:
I installed a markdown viewer plugin in my browser and visited
   file:///home/csilvers/khan/style-guides/style/react.md
and saw that the page rendered in a way that looked good, at least to
me.

Reviewers: jared, emily

Differential Revision: https://phabricator.khanacademy.org/D22265