Add support for a table of contents to the react style guide.
This was a manual process using gh-md-toc, downloaded from here:
   https://raw.githubusercontent.com/ekalinin/github-markdown-toc/master/gh-md-toc
as described here:
   https://github.com/ekalinin/github-markdown-toc

This has a downside that it has to be manually updated every time a
new section is added to the doc, but I think the utility of a TOC
makes it worth it.

Auditors: jared

Test Plan:
Viewed
   file:///home/csilvers/khan/style-guides/style/react.md
in a browser with a markdown-viewer plugin, and saw it look beautiful.