Add "Spaces" section to our JS style guide
Summary:
A quick poll in #web-frontend showed that we favored adding a rule for
whitespace-padded braces to the style-guide. As a result of [the
Braces spaces survey](https://docs.google.com/a/khanacademy.org/forms/d/1lN3-LfXKTlj6kf_-Cv50DNssxrCTF47kGLGQvhi8Bw0/),
we are standardizing on no spaces.

This commit adds a section to our style guide expanding on the new rule.

Test Plan: n/a

Reviewers: john

Reviewed By: john

Differential Revision: https://phabricator.khanacademy.org/D28565